Updates install file

+++ install.sh
@@ -4,43 +4,46 @@ lowercase(){
 }
 
 OS=`lowercase \`uname\``
-echo 'name:'
-read name
-echo 'email:'
-read email
+# echo 'name:'
+# read name
+# echo 'email:'
+# read email
 
-set -v
-rm -rf ~/.ackrc
-rm -rf ~/.bash_profile
-rm -rf ~/.bashrc
-rm -rf ~/.gemrc
-rm -rf ~/.gitconfig
-rm -rf ~/.irbrc
-rm -rf ~/.zshrc
-rm -rf  ~/bin
+configFiles=(
+  "ackrc"
+  "bash_profile"
+  "bashrc"
+  "gemrc"
+  "gitconfig"
+  "irbrc"
+  "zshrc"
+)
 
+for t in ${configFiles[@]}; do
+ echo "Removing .$t"
+ rm -rf ~/.$t
+ echo "symlinking $t to ~/.$t"
+ ln -s $PWD/$t ~/.$t
+done
+
+echo "Creating dotfile-local directory for non-repo config files"
 mkdir -p "$PWD"/dotfiles-local
+[ -e ~/dotfiles-local ] || ln -s "$PWD"/dotfiles-local ~/dotfiles-local
 
-ln -s "$PWD"/ackrc ~/.ackrc
-ln -s "$PWD"/bash_profile ~/.bash_profile
-ln -s "$PWD"/bashrc ~/.bashrc
-ln -s "$PWD"/gitconfig ~/.gitconfig
-ln -s "$PWD"/gemrc ~/.gemrc
-ln -s "$PWD"/irbrc ~/.irbrc
-ln -s "$PWD"/zshrc ~/.zshrc
+rm -rf ~/bin
 ln -s "$PWD"/bin ~/bin
-ln -s "$PWD"/dotfiles-local ~/dotfiles-local
 
+echo "Installing Xcode theme"
 XCODE_THEME_INSTALL_PATH=~/Library/Developer/Xcode/UserData/FontAndColorThemes
 XCODE_THEME_NAME=base16-twilight.dark.xccolortheme
 
 [ -e $XCODE_THEME_INSTALL_PATH ] || mkdir -p $XCODE_THEME_INSTALL_PATH
-ln -s "$PWD"/macOS/"$XCODE_THEME_NAME" "$XCODE_THEME_INSTALL_PATH"/"$XCODE_THEME_NAME"
+[ -e "$XCODE_THEME_INSTALL_PATH"/"$XCODE_THEME_NAME" ] || ln -s "$PWD"/macOS/"$XCODE_THEME_NAME" "$XCODE_THEME_INSTALL_PATH"/"$XCODE_THEME_NAME"
 
+echo "Updating Homebrew"
 brew bundle install
 
-set +v
-
+echo "Updating git config.loca with user and email"
 printf "%s[user]%s\n  name = %s$name%s\n  email = $email" >> ~/dotfiles-local/gitconfig.local
 if [ "{$OS}"=="darwin" ]
 then
